Created a meteorology sub-folder and moved meteorology and meteorology_windfield into it. Moved SiteLayout class over to its own module in the dispersion_model folder. Updated imports and __init__ files accordingly.
