+++ src/pyelq/dispersion_model/gaussian_plume.py
@@ -19,7 +19,7 @@ import numpy as np
 from pyelq.coordinate_system import ENU, LLA
 from pyelq.dispersion_model.dispersion_model import DispersionModel
 from pyelq.gas_species import GasSpecies
-from pyelq.meteorology import Meteorology, MeteorologyGroup
+from pyelq.meteorology.meteorology import Meteorology, MeteorologyGroup
 from pyelq.sensor.beam import Beam
 from pyelq.sensor.satellite import Satellite
 from pyelq.sensor.sensor import Sensor, SensorGroup
