+++ src/pyelq/preprocessing.py
@@ -10,7 +10,7 @@ from typing import Union
 import numpy as np
 import pandas as pd
 
-from pyelq.meteorology import Meteorology, MeteorologyGroup
+from pyelq.meteorology.meteorology import Meteorology, MeteorologyGroup
 from pyelq.sensor.sensor import Sensor, SensorGroup
 from pyelq.support_functions.spatio_temporal_interpolation import temporal_resampling
